Auth API implementation

+++ server/routes/category.py
@@ -5,7 +5,7 @@ from mongoengine import *
 from schema.category import *
 from model import Category
 
-
+from routes.auth import auth
 # Connect to mongodb
 connect('saggezza_db', host='localhost', port=27017)
 
@@ -14,7 +14,13 @@ class CategoryListAPI(Resource):
     # |- /category
     # |- POST: Create a new category
     # \- GET: Return all categories
+
+    @auth.login_required
     def post(self):
+        caller = get_caller(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Must be an admin to create a new category", "error"), 400
+
         req = parse(request)
         errors = CategorySchema().validate(req)
         if errors:
@@ -25,7 +31,11 @@ class CategoryListAPI(Resource):
         category.save()
         return res('Category created successfully', 'success', category=convert_query(category))
 
+    @auth.login_required
     def get(self):
+        caller = get_bearer(request)
+        if caller["role"] != "admin" or caller["role"] != "employee":
+            return res("⛔️ Must be an admin or an employee to get a list of categories", "error"), 400
         categories = Category.objects().all()
         return res('All categories returned', 'success', categories=convert_query(categories))
 
@@ -36,7 +46,11 @@ class CategoryAPI(Resource):
     # |- GET: Return category
     # \- DELETE: Delete category
 
+    @auth.login_required
     def put(self, id):
+        caller = get_caller(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Must be an admin to update a category", "error"), 400
         try:
             req = parse(request)
             category = Category.objects(id=id)
@@ -48,14 +62,21 @@ class CategoryAPI(Resource):
         except:
             return res("Category doesn't exist", 'error'), 400
 
+    @auth.login_required
     def get(self, id):
+
         try:
             category = Category.objects(id=id)[0]
             return res('Retrieved Successfully', 'success', category=convert_query(category))
         except:
             return res("Category doesn't exist", 'error'), 400
 
+    @auth.login_required
     def delete(self, id):
+        caller = get_caller(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Must be an admin delete a category", "error"), 400
+
         try:
             category = Category.objects(id=id)
             category.delete()
